requirements.txt
spacy==2.3.0
python-dotenv==0.15.0
beautifulsoup4==4.9.3
selenium==3.141.0
requests==2.25.1
